+++ db/migrate/20240108113131_add_date_received_to_conversion_grant_payment_certificate_task.rb
@@ -1,0 +1,5 @@
+class AddDateReceivedToConversionGrantPaymentCertificateTask < ActiveRecord::Migration[7.0]
+  def change
+    add_column :conversion_tasks_data, :receive_grant_payment_certificate_date_received, :date
+  end
+end